+++ R/countries.R
@@ -1,13 +1,16 @@
 #' Word countries ISO code.
 #'
 #' \itemize{
-#'   \item CODE  ISO 3 characters codes
-#'   \item COUNTRY Full countries names
+#'   \item code  ISO 3 characters codes
+#'   \item country Full countries names
+#'   \item subcode Countries subcodes
+#'   \item subname Countries subname
 #' }
 #' 
 #' @docType data
 #' @keywords datasets
 #' @name countries
 #' @usage data("countries")
-#' @format A data frame with 244 rows and 2 variables
+#' @format A data frame with 4707 rows and 4 variables
+#' @references \url{http://www.unece.org/cefact/codesfortrade/codes_index.html}
 NULL