+++ NEWS.md
@@ -1,5 +1,7 @@
 # grtrends 1.3.2 (unreleased)
 
+- Added support for sub-countries (#25). Ex.: `gtrends("NHL", geo = "CA-QC")` will return trends data for Québec province in Canada. The list of supported sub-countries can be obtained via `data(countries)`.
+
 - Data parsing should work for any data returned by Google Trends (i.e. countries independent).
 
 - Better support for queries using keywords in different languages (#50, #57). Ex.: `gtrends("蘋果", geo = "TW")`